The most significant changes include the addition of a new model `QSM_M` to the `MY_MODELS` enumeration in `hamil.h` and the corresponding `DECL_ENUM_ELEMENT(QSM_M)` to the `BEGIN_ENUM(MY_MODELS)` block. The `Hamiltonian` class constructors were also modified to include additional parameters and some constructors were removed. The `prettyPrint` function was updated to accept a templated vector type. The `hamiltonian` function was updated to use pre-increment (`++k`) instead of post-increment (`k++`) in its loops for efficiency. The `buildHamiltonian` function was moved from the `protected` to the `public` section of the `Hamiltonian` class.
In `XYZ.h`, the `locEnergy` function was updated to use pre-increment (`++k`) instead of post-increment (`k++`) in its loops for efficiency. The `hamiltonian` function was removed. In `SYK2.h`, the `hamiltonian` function was updated to use a templated version of the `GOE` function from the `randomGen` class.

In `nqs.h`, changes include modifications to the formatting of the code and the addition of new methods for training the neural quantum state, collecting data, and other operations. The changes in the `public:` section of `nqs.h` file seem to be related to the restructuring of the code and the addition of new methods.

In `user_interface.h`, new model includes were added, the `ModP` struct was refactored to include new parameters for different models, and the `setDefault()` function was refactored to set default parameters for different models. A new function `defineHilbert()` was added to define Hilbert space, and the `defineModel()` function was refactored to use `defineHilbert()` function. A new function `checkETH()` was added to check ETH property of a model, and `hamiltonian()` function calls were changed to `buildHamiltonian()`. A new function `quadraticStatesManifold()` was added to compute states manifold for quadratic models. The `parseModel()` function in `user_interface.cpp` was refactored to parse parameters for different models. The subproject commit in `source` was updated.

In `quantum_sun.h`, the code changes are primarily additions to the file, which appears to define a class `QSM` that represents a quantum system model. The class includes various properties of the quantum system, such as the number of particles, the Hamiltonian of the system, and various parameters related to the system's state. The class also includes methods for initializing the system, calculating the Hamiltonian, and other operations.

In the namespace Hilbert in hilbert.h file, several new type aliases have been added for better code readability and maintainability. These include GSymV, SymOpV, SymGV, RPair, and RPairV. The destructor and constructors of the HilbertSpace class have been updated. The destructor now includes a call to the DESTRUCTOR_CALL macro. The default constructor and the constructor that takes a shared_ptr to a Lattice object as a parameter have been updated to call the 'init' and 'hi' methods. The copy constructor and the move constructor have been updated to use the std::move function to move the values of the members of the source object to the new object. New methods 'hi' and 'init' have been added to the HilbertSpace class. The 'hi' method logs information about the Hilbert space, and the 'init' method initializes the members of the HilbertSpace object. New methods 'checkGSym', 'checkGVal', 'checkU1', and 'checkU1Val' have been added to the HilbertSpace class. These methods check for the presence of certain global symmetries and their values.

In the namespace Operators in nqs_operator.h file, the code changes are related to the type casting of a variable. The variable 'N' was previously cast to 'double' type, but it has been changed to 'long double' type. In the namespace NQSAv in nqs_operator.h file, the log message in the destructor of the MeasurementNQS class has been updated to provide more specific information. The message now includes "of the NQS averages." at the end. The indentation of the constructor parameters in the MeasurementNQS class has been adjusted for better readability. The error messages in the catch handlers of the normalize and measure methods in the MeasurementNQS class have been updated to provide more specific information about the type of operators that may have caused the problem.